Lesson

+++ Main.java
@@ -3,9 +3,10 @@ public class Main {
     // static int age;
 
     public static void main(String[] arguments) {
-        int number1 = 5;
-        double number2 = number1;
+        double number1 = 5;
+        int number2 = (int) number1;
 
         System.out.println(number2);
     }
 }
+
